Revert "feat(chatwitheino): add middleware test stack"
This reverts commit 17ecc4b6081a6907826504c05824eeccdf7db091.
