--- quickstart/chatwitheino/AGENTS.md
@@ -1,10 +1,0 @@
-# ChatWithEino Middleware Test Notes
-
-This file is intentionally small and exists so the `agentsmd` middleware has a
-default document to inject during local ChatWithEino runs.
-
-- Prefer concise answers when explaining Eino concepts.
-- When code context is needed, use the available filesystem tools before
-  answering.
-- If the user asks about middleware behavior, mention which middleware or tool
-  was used when that is observable.
