+++ quickstart/chatwitheino/server/server_test.go
@@ -93,22 +93,6 @@ func slowAgent(delay time.Duration, reply string) *mockAgent {
 	}
 }
 
-func interruptingAgent() *mockAgent {
-	return &mockAgent{
-		name: "interrupting-agent",
-		onRun: func(ctx context.Context, _ *adk.TypedAgentInput[*schema.Message], gen *adk.AsyncGenerator[*adk.TypedAgentEvent[*schema.Message]]) {
-			defer gen.Close()
-			gen.Send(adk.TypedStatefulInterrupt[*schema.Message](ctx,
-				&commontool.ApprovalInfo{
-					ToolName:        "answer_from_document",
-					ArgumentsInJSON: `{"question":"test"}`,
-				},
-				`{"question":"test"}`,
-			))
-		},
-	}
-}
-
 // ---------------------------------------------------------------------------
 // helper: setup test server + HTTP client
 // ---------------------------------------------------------------------------
@@ -247,100 +231,6 @@ func TestChatNormalFlow(t *testing.T) {
 	}
 	if !found {
 		t.Error("assistant message not found in session history")
-	}
-}
-
-func TestInitialLoopPersistsCheckpointForApprovalResume(t *testing.T) {
-	agent := interruptingAgent()
-	srv, _, cleanup := newTestServer(t, agent)
-	defer cleanup()
-
-	sessionID := createSession(t, srv)
-	sess, _ := srv.cfg.Store.GetOrCreate(sessionID)
-
-	ts := srv.getTurnState(sessionID)
-	ts.mu.Lock()
-	loop := srv.newLoop(sess, sessionID, false)
-	ts.loop = loop
-	ts.iterReady = make(chan iterEnvelope[*schema.Message], 1)
-	ts.iterDone = make(chan iterResult[*schema.Message], 1)
-	ts.mu.Unlock()
-
-	loop.Push(&ChatItem{Query: "needs approval"})
-	loop.Run(context.Background())
-
-	var envelope iterEnvelope[*schema.Message]
-	select {
-	case envelope = <-ts.iterReady:
-	case <-time.After(5 * time.Second):
-		t.Fatal("timeout waiting for iterReady")
-	}
-
-	var interruptID string
-	for {
-		event, ok := envelope.events.Next()
-		if !ok {
-			break
-		}
-		if event == nil || event.Action == nil || event.Action.Interrupted == nil {
-			continue
-		}
-		for _, ic := range event.Action.Interrupted.InterruptContexts {
-			if ic.IsRootCause {
-				interruptID = ic.ID
-				break
-			}
-		}
-	}
-	if interruptID == "" {
-		t.Fatal("expected interrupt ID")
-	}
-
-	envelope.done <- iterResult[*schema.Message]{interruptID: interruptID}
-
-	result := loop.Wait()
-	if result.ExitReason == nil {
-		t.Fatal("expected interrupted loop exit reason")
-	}
-	if !result.CheckpointAttempted {
-		t.Fatal("expected initial interrupted loop to save checkpoint")
-	}
-	if result.CheckpointErr != nil {
-		t.Fatalf("checkpoint error: %v", result.CheckpointErr)
-	}
-	if _, ok, err := srv.cfg.CheckPointStore.Get(context.Background(), sessionID); err != nil {
-		t.Fatalf("checkpoint get error: %v", err)
-	} else if !ok {
-		t.Fatal("expected checkpoint to exist for approval resume")
-	}
-}
-
-func TestCheckpointStoreDeleteHidesCheckpoint(t *testing.T) {
-	agent := simpleReplyAgent("test")
-	srv, _, cleanup := newTestServer(t, agent)
-	defer cleanup()
-
-	store := srv.cfg.CheckPointStore
-	if err := store.Set(context.Background(), "session-1", []byte("checkpoint")); err != nil {
-		t.Fatalf("set checkpoint: %v", err)
-	}
-	if _, ok, err := store.Get(context.Background(), "session-1"); err != nil {
-		t.Fatalf("get checkpoint: %v", err)
-	} else if !ok {
-		t.Fatal("expected checkpoint before delete")
-	}
-
-	deleter, ok := store.(adk.CheckPointDeleter)
-	if !ok {
-		t.Fatal("checkpoint store should support delete")
-	}
-	if err := deleter.Delete(context.Background(), "session-1"); err != nil {
-		t.Fatalf("delete checkpoint: %v", err)
-	}
-	if _, ok, err := store.Get(context.Background(), "session-1"); err != nil {
-		t.Fatalf("get after delete: %v", err)
-	} else if ok {
-		t.Fatal("expected checkpoint to be hidden after delete")
 	}
 }
